models/requirements.txt
fastapi
tensorflow
uvicorn
Pillow
requests